Move config classes to a package
