+++ .gitignore
@@ -8,6 +8,7 @@ __pycache__/*
 metastore_db/*
 spark-warehouse/*
 derby.log
+conf/__pycache__/*
 tests/__pycache__/*
 schema/__pycache__/*
 extract/__pycache__/*
